agent-service/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
pydantic==2.11.10
python-dotenv==1.1.1
httpx==0.28.1
crewai[google-genai]==1.10.1
crewai-tools==1.10.1
anthropic==0.40.0
google-generativeai==0.8.5
google-genai